Adding pagination and more minor functions

+++ admin/functions.php
@@ -53,7 +53,7 @@ function ShowAllCategories() {
                                      echo "<tr>";
                                      echo "<td>{$cat_id}</td>";
                                      echo "<td>{$cat_title}</td>";
-                                     echo "<td><a href='categories.php?delete={$cat_id}'>Delete</a></td>";
+                                     echo "<td><a onClick=\"javascript: return confirm('Are you sure you want to delete?'); \" href='categories.php?delete={$cat_id}'>Delete</a></td>";
                                      echo "<td><a href='categories.php?edit={$cat_id}'>Edit</a></td>";
                                      echo "</tr>";
